VariantMap now displays gap sizes or an axis to make it clear that the display is non-linear in base space. Fixed null pointer when creating project from genome. Fixed invalid argument when accessing a cache file larger than 2GB.
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@3032 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/util/CachedSeekableStream.java
@@ -93,7 +93,7 @@ public class CachedSeekableStream extends SeekableStream {
 
     private void cachedSeek(int blockOffset, int positionOffset) throws IOException{
         //from offset, calculate actual position in file
-        int actualOffset = (numBlocks * 4) + (blockOffset -1) * bufferSize;
+        long actualOffset = (numBlocks * 4L) + (blockOffset - 1L) * bufferSize;
 
         //retrieve from stream
         FileInputStream cacheStream = new FileInputStream(cacheFile);
